+++ apps/web/app/search/mockup.ts
